+++ Products/Archetypes/Layer.py
@@ -2,11 +2,11 @@ from Products.Archetypes.interfaces.layer import ILayerContainer
 from ExtensionClass import Base
 from App.class_init import InitializeClass
 from AccessControl import ClassSecurityInfo
-from zope.interface import implements
+from zope.interface import implementer
 
 
+@implementer(ILayerContainer)
 class DefaultLayerContainer(Base):
-    implements(ILayerContainer)
 
     security = ClassSecurityInfo()